refactor core redux store code out of index into src/store/store + update imports
