+++ src/hooks/useDispatch.ts
@@ -1,8 +1,7 @@
 // eslint-disable-next-line no-restricted-imports
 import { useDispatch as _useDispatch } from "react-redux";
 import { ThunkDispatch } from "redux-thunk";
-import { RootActions } from "store/actions";
-import { RootState } from "index";
+import { RootState, RootActions } from "store";
 
 export const useDispatch = () =>
   _useDispatch<ThunkDispatch<RootState, void, RootActions>>();
