+++ src/hooks/useSelector.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any*/
 // eslint-disable-next-line no-restricted-imports
 import { TypedUseSelectorHook, useSelector as _useSelector } from "react-redux";
-import { RootState } from "index";
+import { RootState } from "store";
 import { WithId } from "utils/id";
 import { useMemo, useRef } from "react";
