+++ src/index.tsx
@@ -9,17 +9,9 @@ import LogRocket from "logrocket";
 // eslint-disable-next-line no-restricted-imports
 import mixpanel from "mixpanel-browser";
 
-import { Provider } from "react-redux";
-import { createStore, combineReducers, applyMiddleware, Reducer } from "redux";
-import thunkMiddleware from "redux-thunk";
-import { createFirestoreInstance, firestoreReducer } from "redux-firestore";
-import {
-  ReactReduxFirebaseProvider,
-  firebaseReducer,
-  isLoaded,
-  FirebaseReducer,
-} from "react-redux-firebase";
-import { composeWithDevTools } from "redux-devtools-extension";
+import { Provider as ReduxStoreProvider } from "react-redux";
+import { createFirestoreInstance } from "redux-firestore";
+import { ReactReduxFirebaseProvider, isLoaded } from "react-redux-firebase";
 
 import firebase from "firebase/app";
 import "firebase/analytics";
@@ -46,13 +38,9 @@ import {
 } from "secrets";
 import { FIREBASE_CONFIG } from "settings";
 
-import { VenueTemplateReducers, MiscReducers } from "store/reducers";
-
 import * as serviceWorker from "./serviceWorker";
 import { activatePolyFills } from "./polyfills";
-
-import { Firestore } from "types/Firestore";
-import { User } from "types/User";
+import { store } from "./store";
 
 import {
   PerformanceTrace,
@@ -107,30 +95,6 @@ const rrfConfig = {
   userProfile: "users",
   useFirestoreForProfile: true,
 };
-
-// Add firebase to reducers
-const rootReducer = combineReducers({
-  firebase: firebaseReducer as Reducer<FirebaseReducer.Reducer<User>>,
-  firestore: firestoreReducer as Reducer<Firestore>,
-  ...VenueTemplateReducers,
-  ...MiscReducers,
-});
-
-export type RootState = ReturnType<typeof rootReducer>;
-
-const initialState = {};
-const store = createStore(
-  rootReducer,
-  initialState,
-  composeWithDevTools(
-    applyMiddleware(
-      thunkMiddleware,
-      LogRocket.reduxMiddleware() // logrocket needs to be last
-    )
-  )
-);
-
-export type AppDispatch = typeof store.dispatch;
 
 const rrfProps = {
   firebase,
@@ -274,7 +238,7 @@ traceReactScheduler("initial render", performance.now(), () => {
       <ThemeProvider theme={theme}>
         <Elements stripe={stripePromise}>
           <DndProvider backend={HTML5Backend}>
-            <Provider store={store}>
+            <ReduxStoreProvider store={store}>
               <ReactReduxFirebaseProvider {...rrfProps}>
                 <AuthIsLoaded>
                   <CustomSoundsProvider
@@ -285,7 +249,7 @@ traceReactScheduler("initial render", performance.now(), () => {
                   </CustomSoundsProvider>
                 </AuthIsLoaded>
               </ReactReduxFirebaseProvider>
-            </Provider>
+            </ReduxStoreProvider>
           </DndProvider>
         </Elements>
       </ThemeProvider>
