+++ src/store/index.ts
@@ -1,0 +1,4 @@
+export type { AppDispatch, RootState } from "./store";
+export { store } from "./store";
+
+export type { RootActions } from "./actions";
