+++ src/types/SparkleSelector.ts
@@ -1,3 +1,3 @@
-import { RootState } from "index";
+import { RootState } from "store";
 
 export type SparkleSelector<T> = (state: RootState) => T;
