+++ src/utils/firestoreSelectors.ts
@@ -1,4 +1,4 @@
-import { RootState } from "index";
+import { RootState } from "store";
 import {
   FirestoreData,
   FirestoreOrdered,
